requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pybaseball>=2.2.7
matplotlib>=3.8.0
scikit-learn>=1.4.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
polars>=1.5.0
ruff>=0.6.0
black>=24.4.0
pytest>=8.0.0
tabulate>=0.9.0
